FMT Debug Print Function
Created FMT debug print function which looks at what core called the function and prints to the appropriate UART instance. This will allow for both cores to print any data without waiting for each other and while allowing the USB com port to be used to communicate with the Python script.