Seemingly Resolved Crashing Issue
After running the debugger with just core 0, it managed to catch a call to _exit which was called after the JSON library failed to parse the input string due to a missing closing }, resulting in an exception and subsequent exit. Still had a crash, and the debugger caught a hard fault irq which was called deep in the call stack of FMTDebug function. As such, all calls to the function have been disabled until it is resolved.